Add files via upload
Take Your Github Repository To The Next Level 🚀️

Web development refers to the creation, building, and maintenance of websites. It includes aspects such as web design, web publishing, web programming, and database management 1. Frontend development and backend development are two ways to classify web development. Frontend development is the part of a website where the user interacts directly and is also referred to as the ‘client-side’ of the application.